Updated edit error
As i separated DELETE and EDIT functionality into two different form tag, so I had to put the input (which contain the particular id of the product) in both of those tag. But I didn't put it inside the EDIT form tag and as a result it was showing error. Fixed that in this commit.